Refactor(jira): Restrict method access to protected for some classes.
